Patch 0522
・AthenaProject Rev.548よりmerge、thanks to ico776さん
	スキルユニットの発動範囲読み込み処理を拡張し、
	発動範囲をスキルレベル毎に設定できるように修正（skill.*）
・0519で、core_ExceptionRoutine() の漏れを修正（core.c）
・不要なキャストを削除（db.c）
・ギルドデータ受信時、トリッキーなmemcpyを使わないように修正（guild.c）
・nullpoチェック追加（skill.c）

+++ src/common/db.c
@@ -762,7 +762,7 @@ struct csvdb_data* csvdb_open(const char* file, int skip_comment)
 	}
 
 	while( fgets( buf, sizeof(buf)-1, fp) ) {
-		int max = 0;
+		size_t max = 0;
 		char *s;
 		struct csvdb_line *line;
 
@@ -775,7 +775,7 @@ struct csvdb_data* csvdb_open(const char* file, int skip_comment)
 			csv->index = (int*)aRealloc(
 				csv->index, sizeof(int) * csv->row_max );
 		}
-		max = (int)strlen(buf);
+		max = strlen(buf);
 		if( buf[max-1] == '\n' )
 			max--;
